+++ src/app/app-list/list/list-detail/list-detail.component.ts
@@ -49,9 +49,10 @@ export class ListDetailComponent implements OnInit {
 
       onSelect(list, itemkey): void {
 
-        this.appService.afs.collection('lists').doc(list.id).update({
-            ['items.'+itemkey]: fs.firestore.FieldValue.delete()
-        })        
+        if (confirm(this.appService.language.m7))
+            this.appService.afs.collection('lists').doc(list.id).update({
+                ['items.'+itemkey]: fs.firestore.FieldValue.delete()
+            })        
           
     }
